Create basic player and placeholder map
Create a controllable player in a separate player.tscn scene.
The root node of the player is a CharacterBody2D with a square CollisionShape2D aligning with the sprites in AnimatedSprite2D.
The current player sprites are temporary, but serve to show the direction of the player when moving.

Create a placeholder map in the primary scene, main_scene.tscn.
The tiles used in the TileMapLayer will be replaced with final ones in the future.

All sprites are located in assets/img/.