Create interactable objects and a dialogue system
Create a new InteractableObject class which allows dialogue, read from a JSON file, to appear on the player's screen.

The InteractableObject class contains the JSON parser for dialogue and requires an Area2D to define its interaction boundaries and a JSON file to define its dialogue. These are specified through exported variables.
When this object is interacted with, the Dialogue class is called, and the player's movement is disabled.

The Dialogue class contains functions to receive dialogue in the form of an Array and handles the display of dialogue on screen.
Two fonts, dialogue_box_normal.tres and dialogue_box_italicized.tres, are located under assets/fonts/ and customize dialogue boxes to use a bold, monospace font with serifs.

The player is able to skip portions of dialogue by pressing the interact button as the dialogue is being displayed.